# docs/swagger.yaml
definitions:
  controller.CreateReportRequest:
    properties:
      body:
        description: 本文
        type: string
      tasks:
        description: 実施したタスクリスト
        items:
          $ref: '#/definitions/entity.Task'
        type: array
    type: object
  controller.InviteUserRequest:
    properties:
      email:
        description: メールアドレス
        type: string
    type: object
  controller.OrganizationsResponse:
    properties:
      organizations:
        description: 組織リスト
        items:
          $ref: '#/definitions/entity.Organization'
        type: array
    type: object
  controller.ReportsResponse:
    properties:
      reports:
        description: 日報リスト
        items:
          $ref: '#/definitions/entity.Report'
        type: array
    type: object
  controller.UpdateOrganizationRequest:
    properties:
      code:
        description: 組織コード
        type: string
      mission:
        description: ミッション
        type: string
      name:
        description: 組織名
        type: string
      value:
        description: バリュー
        type: string
      vision:
        description: ビジョン
        type: string
    type: object
  controller.UsersResponse:
    properties:
      users:
        description: ユーザリスト
        items:
          $ref: '#/definitions/entity.User'
        type: array
    type: object
  entity.ErrorResponse:
    properties:
      message:
        type: string
    type: object
  entity.Mvv:
    properties:
      mission:
        description: ミッション
        type: string
      value:
        description: バリュー
        type: string
      vision:
        description: ビジョン
        type: string
    type: object
  entity.Organization:
    properties:
      code:
        description: 組織コード
        type: string
      id:
        description: 組織ID
        type: string
      mvv:
        $ref: '#/definitions/entity.Mvv'
      name:
        description: 組織名
        type: string
    type: object
  entity.Report:
    properties:
      body:
        description: 本文
        type: string
      id:
        description: 日報レスポンス
        type: string
      reviewBody:
        description: レビュー本文
        type: string
      tasks:
        description: 実施したタスクリスト
        items:
          $ref: '#/definitions/entity.Task'
        type: array
      userId:
        description: ユーザID
        type: string
    type: object
  entity.Task:
    properties:
      finishedAt:
        description: 終了日時
        type: string
      name:
        description: タスク名
        type: string
      startedAt:
        description: 開始日時
        type: string
    type: object
  entity.User:
    properties:
      email:
        description: メールアドレス
        type: string
      id:
        description: ユーザID
        type: string
      name:
        description: ユーザ名
        type: string
      organizations:
        description: 所属する組織リスト
        items:
          $ref: '#/definitions/entity.UserOrganization'
        type: array
    type: object
  entity.UserOrganization:
    properties:
      id:
        description: 組織ID
        type: string
      is_admin:
        description: ロール
        type: boolean
    type: object
info:
  contact: {}
paths:
  /organizations:
    get:
      consumes:
      - application/json
      description: 自分が所属する組織のみ取得できる
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.OrganizationsResponse'
        "400":
          description: BadRequest
          schema:
            $ref: '#/definitions/entity.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/entity.ErrorResponse'
      security:
      - Bearer: []
      summary: 組織リスト取得API
      tags:
      - Organization
  /organizations/{organizationCode}:
    get:
      consumes:
      - application/json
      parameters:
      - description: 組織コード
        in: path
        name: organizationCode
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Organization'
        "400":
          description: BadRequest
          schema:
            $ref: '#/definitions/entity.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/entity.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/entity.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/entity.ErrorResponse'
      security:
      - Bearer: []
      summary: 組織取得API
      tags:
      - Organization
    put:
      consumes:
      - application/json
      parameters:
      - description: 組織コード
        in: path
        name: organizationCode
        required: true
        type: string
      - description: 組織更新リクエスト
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.UpdateOrganizationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Organization'
        "400":
          description: BadRequest
          schema:
            $ref: '#/definitions/entity.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/entity.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/entity.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/entity.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/entity.ErrorResponse'
      security:
      - Bearer: []
      summary: 組織更新API
      tags:
      - Organization
  /organizations/{organizationCode}/reports:
    get:
      consumes:
      - application/json
      parameters:
      - description: ユーザID
        in: query
        name: userId
        required: true
        type: string
      - description: 組織コード
        in: path
        name: organizationCode
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.ReportsResponse'
        "400":
          description: BadRequest
          schema:
            $ref: '#/definitions/entity.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/entity.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/entity.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/entity.ErrorResponse'
      security:
      - Bearer: []
      summary: 日報リスト取得API
      tags:
      - Report
    post:
      consumes:
      - application/json
      parameters:
      - description: ユーザID
        in: query
        name: userId
        required: true
        type: string
      - description: 組織コード
        in: path
        name: organizationCode
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/controller.CreateReportRequest'
        "400":
          description: BadRequest
          schema:
            $ref: '#/definitions/entity.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/entity.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/entity.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/entity.ErrorResponse'
      security:
      - Bearer: []
      summary: 日報作成API
      tags:
      - Report
  /organizations/{organizationCode}/reports/{reportId}:
    get:
      consumes:
      - application/json
      parameters:
      - description: ユーザID
        in: query
        name: userId
        required: true
        type: string
      - description: 組織コード
        in: path
        name: organizationCode
        required: true
        type: string
      - description: 日報ID
        in: path
        name: organizationCode
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Report'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/entity.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/entity.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/entity.ErrorResponse'
      security:
      - Bearer: []
      summary: 日報取得API
      tags:
      - Report
  /organizations/{organizationCode}/users:
    get:
      consumes:
      - application/json
      parameters:
      - description: 組織コード
        in: path
        name: organizationCode
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.UsersResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/entity.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/entity.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/entity.ErrorResponse'
      security:
      - Bearer: []
      summary: 組織のユーザリスト取得API
      tags:
      - User
    post:
      consumes:
      - application/json
      parameters:
      - description: 組織コード
        in: path
        name: organizationCode
        required: true
        type: string
      - description: メンバー招待リクエスト
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.InviteUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.User'
        "400":
          description: BadRequest
          schema:
            $ref: '#/definitions/entity.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/entity.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/entity.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/entity.ErrorResponse'
      security:
      - Bearer: []
      summary: メンバー招待API
      tags:
      - User
  /users/me:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.User'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/entity.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/entity.ErrorResponse'
      security:
      - Bearer: []
      summary: ログインユーザー取得API
      tags:
      - User
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"